:bug: recycle-bin.exe / trash not found (Closes: 76)
